Regenerate PDFs with everyone's changes.
An unfortunnate consequence of having the PDFs in the repository is that
when someone submits a PR that contains a PDF with their changes, it's
usually difficult to merge that PDF with the existing one. This problem
can cause the .tex documents to not match the PDF. I'll try to look more
into suggstions mentioned in #6 about using github pages or some other
solution so this doesn't happen.