Remove optional definitions from the diffeq text
The labels were not particularly useful and added extra complecity to
the document.
This also brings diffeq's main.tex closer in line to the main.tex for
the other doucments.
This helps in the eventual goal to extract some of the commonalities of
each document's main.tex into a separate file so that style and setup is
uniform across all the documents.