Credit contributors with script.
Relevant to #3, I have added a file called contributors.tex that thanks
everyone who has contributed to the project on Github. contributors.tex
is generated by make_contributors.py. It uses the GitHub REST API to
get each contributor's name and username on Github.

In the future, I hope to have this contributors.py script automate
updating the contributors.tex file. Ideally, after each push/merge into
the master branch, the contributors.py script would run and update
contributors.tex.

It'd be nice if this could be tied into another action like the
[tectonic typesetting
system](https://tectonic-typesetting.github.io/en-US/), which would
allow the creation of PDFs to be automated. This would allow #6 to be
solved in a much more elegant way if PDFs could still be gitignored but
also generated for everyone to see.