Add clean_quotes script
This should close #38. However, I didn't add the warning about backwards
quotes before running the script, so it's possible that some backwards
quotes got transformed, which will look weird in the tex, likely with a
`` or '' being at the beginning of a line. If you see any occourences of
this, please try to fix it.