Specify font by more generic filename
I originally used the filename that I had see when downloading the
openserif font. However, when this didn't work, I went for a more
generic name, similar to what was shown in the tectonic issue.

If this doesn't work, I may try including the font file directly with
the project.